molesec-attack-service/requirements.dev.txt
APScheduler==3.10.1
Django==4.1.7
Pillow==9.4.0
black==23.3.0
cachetools==5.2.0
django-cors-headers==3.14.0
django-extensions==3.2.1
django-sendgrid-v5==1.2.2
django-stubs==1.16.0
djangorestframework-api-key==2.3.0
djangorestframework-stubs==1.10.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
eventlet==0.30.2
flake8==6.0.0
google-api-core==2.11.0
google-api-python-client==2.84.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.17.2
google==3.0.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
ipython==8.9.0
jupyterlab==3.5.3
isort==5.11.5
langchain==0.0.126
mypy==1.1.1
openai==0.27.2
pre-commit==3.2.2
psycopg2-binary==2.9.5
pydantic[email]==1.10.7
pylint==2.17.0
python-dotenv==1.0.0
requests==2.28.2
sentry-sdk==1.19.1
slack-sdk==3.20.2
tzlocal==2.1 # https://stackoverflow.com/questions/69776414/pytzusagewarning-the-zone-attribute-is-specific-to-pytzs-interface-please-mig
whitenoise==6.3.0